refactor: move backup and restore wallets
