+++ packages/mask/src/extension/popups/pages/Wallet/AddDeriveWallet/index.tsx
@@ -2,11 +2,10 @@ import { memo, useCallback, useRef, useState } from 'react'
 import { useNavigate, useLocation } from 'react-router-dom'
 import { TableContainer, TablePagination, tablePaginationClasses, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
-import { HD_PATH_WITHOUT_INDEX_ETHEREUM } from '@masknet/plugin-wallet'
 import { useAsync, useAsyncFn } from 'react-use'
 import { WalletRPC } from '../../../../../plugins/Wallet/messages'
 import { DeriveWalletTable } from '../components/DeriveWalletTable'
-import { currySameAddress, NetworkPluginID } from '@masknet/web3-shared-base'
+import { currySameAddress, NetworkPluginID, HD_PATH_WITHOUT_INDEX_ETHEREUM } from '@masknet/web3-shared-base'
 import { useWallets } from '@masknet/plugin-infra/web3'
 import { useI18N } from '../../../../../utils'
 import { LoadingButton } from '@mui/lab'
