+++ packages/mask/src/plugins/Wallet/constants.ts
@@ -1,1 +1,1 @@
-export { PLUGIN_ID, HD_PATH_WITHOUT_INDEX_ETHEREUM, UPDATE_CHAIN_STATE_DELAY } from '@masknet/plugin-wallet'
+export { PLUGIN_ID, UPDATE_CHAIN_STATE_DELAY } from '@masknet/plugin-wallet'
