+++ packages/mask/src/plugins/Wallet/services/wallet/index.ts
@@ -3,10 +3,10 @@ import { first, last, omit } from 'lodash-unified'
 import { hexToBytes, toHex } from 'web3-utils'
 import { toBuffer } from 'ethereumjs-util'
 import { personalSign, signTypedData as signTypedData_, SignTypedDataVersion } from '@metamask/eth-sig-util'
-import { isSameAddress, Wallet } from '@masknet/web3-shared-base'
+import { isSameAddress, Wallet, HD_PATH_WITHOUT_INDEX_ETHEREUM } from '@masknet/web3-shared-base'
 import { createPayload, EthereumMethodType, formatEthereumAddress, Transaction } from '@masknet/web3-shared-evm'
 import { api } from '@dimensiondev/mask-wallet-core/proto'
-import { MAX_DERIVE_COUNT, HD_PATH_WITHOUT_INDEX_ETHEREUM } from '@masknet/plugin-wallet'
+import { MAX_DERIVE_COUNT } from '@masknet/plugin-wallet'
 import * as database from './database'
 import * as password from './password'
 import * as Mask from '../maskwallet'
