+++ packages/web3-shared/base/src/utils/index.ts
@@ -11,3 +11,4 @@ export * from './function'
 export * from './percentage'
 export * from './supply'
 export * from './locale'
+export const HD_PATH_WITHOUT_INDEX_ETHEREUM = "m/44'/60'/0'/0"
